Municipality geometries

A simple feature containing the URIs, municipality codes and geometries of municipalities in time. The starting point and end point of each municipality can be determined by combining polygons1909_2009 with another dataset that contains such information.

Usage

polygons1909_2009

Format

A simple feature with 860 rows and 3 variables:

x

Universal Resource Identifier (URI) for each municipality instance in time. For example: http://www.yso.fi/onto/sapo/Maalahti(1908-1972)

kunta_nro

Municipality code, a unique 3-digit code (001-999) assigned for each municipality that stays the same as long as the municipality exists. For example: "475"

geometry

A single list column with geometries

Source

Original data downloaded from ONKI.fi website on 04 Aug 2022: http://onki.fi/en/browser/overview/sapo Data attribution: FinnONTO Consortium: https://seco.cs.aalto.fi/projects/finnonto/