Algorithmic Tools for Open Data in Finland

The sorvi package hosts various functions that are mainly helpful in rOpenGov package maintenance, package authoring and drawing graphs for presentations. Additionally it has some functions that do not (yet) have their own package but are useful in some contexts.

References

See citation("sorvi") https://github.com/rOpenGov/sorvi

Author

Leo Lahti, Juuso Parkkinen, Jussi Paananen, Joona Lehtomaki, Einari Happonen, Juuso Haapanen, and Pyry Kantanen louhos@googlegroups.com

Examples

library(sorvi)